refactor: Move publish helper functions to internal package (#15696)
#### Summary

Some refactoring to make it easier to support Docker publishing. This PR moves some functions to internal helper packages `publish` and `hub` instead of having all the code under the top level commands implementation.

<!--
