+++ cli/cmd/plugin_publish.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
@@ -15,6 +14,8 @@ import (
 
 	cloudquery_api "github.com/cloudquery/cloudquery-api-go"
 	"github.com/cloudquery/cloudquery-api-go/auth"
+	"github.com/cloudquery/cloudquery/cli/internal/hub"
+	"github.com/cloudquery/cloudquery/cli/internal/publish"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 )
@@ -58,24 +59,6 @@ func newCmdPluginPublish() *cobra.Command {
 	return cmd
 }
 
-type PackageJSONV1 struct {
-	Team             string                    `json:"team"`
-	Name             string                    `json:"name"`
-	Message          string                    `json:"message"`
-	Version          string                    `json:"version"`
-	Kind             cloudquery_api.PluginKind `json:"kind"`
-	Protocols        []int                     `json:"protocols"`
-	SupportedTargets []TargetBuild             `json:"supported_targets"`
-	PackageType      string                    `json:"package_type"`
-}
-
-type TargetBuild struct {
-	OS       string `json:"os"`
-	Arch     string `json:"arch"`
-	Path     string `json:"path"`
-	Checksum string `json:"checksum"`
-}
-
 func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) error {
 	tc := auth.NewTokenClient()
 	token, err := tc.GetToken()
@@ -84,7 +67,7 @@ func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) er
 	}
 
 	distDir := cmd.Flag("dist-dir").Value.String()
-	pkgJSON, err := readPackageJSON(distDir)
+	pkgJSON, err := publish.ReadPackageJSON(distDir)
 	if err != nil {
 		return fmt.Errorf("failed to read package.json: %w", err)
 	}
@@ -115,7 +98,7 @@ func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) er
 	}
 
 	// create new draft version
-	err = createNewPluginDraftVersion(ctx, c, teamName, pluginName, pkgJSON)
+	err = publish.CreateNewPluginDraftVersion(ctx, c, teamName, pluginName, pkgJSON)
 	if err != nil {
 		return fmt.Errorf("failed to create new draft version: %w", err)
 	}
@@ -124,7 +107,7 @@ func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) er
 		// upload table schemas
 		fmt.Println("Uploading table schemas...")
 		tablesJSONPath := filepath.Join(distDir, "tables.json")
-		err = uploadTableSchemas(ctx, c, teamName, pluginName, tablesJSONPath, pkgJSON)
+		err = publish.UploadTableSchemas(ctx, c, teamName, pluginName, tablesJSONPath, pkgJSON)
 		if err != nil {
 			return fmt.Errorf("failed to upload table schemas: %w", err)
 		}
@@ -133,7 +116,7 @@ func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) er
 	// upload docs
 	fmt.Println("Uploading docs...")
 	docsDir := filepath.Join(distDir, "docs")
-	if err := uploadDocs(ctx, c, teamName, string(pkgJSON.Kind), pluginName, pkgJSON.Version, docsDir, true); err != nil {
+	if err := publish.UploadPluginDocs(ctx, c, teamName, string(pkgJSON.Kind), pluginName, pkgJSON.Version, docsDir, true); err != nil {
 		return fmt.Errorf("failed to upload docs: %w", err)
 	}
 
@@ -141,7 +124,7 @@ func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) er
 	fmt.Println("Uploading binaries...")
 	for _, t := range pkgJSON.SupportedTargets {
 		fmt.Printf("- Uploading %s_%s...\n", t.OS, t.Arch)
-		err = uploadPluginBinary(ctx, c, teamName, pluginName, t.OS, t.Arch, path.Join(distDir, t.Path), pkgJSON)
+		err = publish.UploadPluginBinary(ctx, c, teamName, pluginName, t.OS, t.Arch, path.Join(distDir, t.Path), pkgJSON)
 		if err != nil {
 			return fmt.Errorf("failed to upload binary: %w", err)
 		}
@@ -163,7 +146,7 @@ func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) er
 			return fmt.Errorf("failed to finalize plugin version: %w", err)
 		}
 		if resp.HTTPResponse.StatusCode > 299 {
-			return errorFromHTTPResponse(resp.HTTPResponse, resp)
+			return hub.ErrorFromHTTPResponse(resp.HTTPResponse, resp)
 		}
 		fmt.Println("Success!")
 		fmt.Printf("%s/%s@%s is now available on the CloudQuery Hub.\n", teamName, pluginName, pkgJSON.Version)
@@ -174,157 +157,4 @@ func runPluginPublish(ctx context.Context, cmd *cobra.Command, args []string) er
 	fmt.Println("\nNote: this plugin version is marked as draft=true. You can preview and finalize it on the CloudQuery Hub, or run `cloudquery plugin publish` with the --finalize flag.")
 
 	return nil
-}
-
-func createNewPluginDraftVersion(ctx context.Context, c *cloudquery_api.ClientWithResponses, teamName, pluginName string, pkgJSON PackageJSONV1) error {
-	targets := make([]string, len(pkgJSON.SupportedTargets))
-	checksums := make([]string, len(pkgJSON.SupportedTargets))
-	for i, t := range pkgJSON.SupportedTargets {
-		targets[i] = fmt.Sprintf("%s_%s", t.OS, t.Arch)
-		checksums[i] = strings.TrimPrefix(t.Checksum, "sha256:")
-	}
-
-	body := cloudquery_api.CreatePluginVersionJSONRequestBody{
-		Message:          pkgJSON.Message,
-		PackageType:      cloudquery_api.CreatePluginVersionJSONBodyPackageType(pkgJSON.PackageType),
-		Protocols:        pkgJSON.Protocols,
-		SupportedTargets: targets,
-		Checksums:        checksums,
-	}
-	resp, err := c.CreatePluginVersionWithResponse(ctx, teamName, pkgJSON.Kind, pluginName, pkgJSON.Version, body)
-	if err != nil {
-		return fmt.Errorf("failed to create plugin version: %w", err)
-	}
-	if resp.HTTPResponse.StatusCode > 299 {
-		err := errorFromHTTPResponse(resp.HTTPResponse, resp)
-		if resp.HTTPResponse.StatusCode == http.StatusForbidden {
-			return fmt.Errorf("%w. Hint: You may need to create the plugin first", err)
-		}
-		return err
-	}
-	return nil
-}
-
-func uploadTableSchemas(ctx context.Context, c *cloudquery_api.ClientWithResponses, teamName, pluginName, tablesJSONPath string, pkgJSON PackageJSONV1) error {
-	b, err := os.ReadFile(tablesJSONPath)
-	if err != nil {
-		return fmt.Errorf("failed to read tables.json: %w", err)
-	}
-	tables := make([]cloudquery_api.PluginTableCreate, 0)
-	err = json.Unmarshal(b, &tables)
-	if err != nil {
-		return fmt.Errorf("failed to parse tables.json: %w", err)
-	}
-	body := cloudquery_api.CreatePluginVersionTablesJSONRequestBody{
-		Tables: tables,
-	}
-	resp, err := c.CreatePluginVersionTablesWithResponse(ctx, teamName, pkgJSON.Kind, pluginName, pkgJSON.Version, body)
-	if err != nil {
-		return fmt.Errorf("failed to upload table schemas: %w", err)
-	}
-	if resp.HTTPResponse.StatusCode > 299 {
-		return errorFromHTTPResponse(resp.HTTPResponse, resp)
-	}
-	return nil
-}
-
-func uploadDocs(ctx context.Context, c *cloudquery_api.ClientWithResponses, teamName, pluginKind, pluginName, version, docsDir string, replace bool) error {
-	dirEntries, err := os.ReadDir(docsDir)
-	if err != nil {
-		return fmt.Errorf("failed to read docs directory: %w", err)
-	}
-
-	pages := make([]cloudquery_api.PluginDocsPageCreate, 0, len(dirEntries))
-	for _, dirEntry := range dirEntries {
-		if dirEntry.IsDir() {
-			continue
-		}
-		fileExt := filepath.Ext(dirEntry.Name())
-		if fileExt != ".md" {
-			continue
-		}
-		content, err := os.ReadFile(filepath.Join(docsDir, dirEntry.Name()))
-		if err != nil {
-			return fmt.Errorf("failed to read docs file: %w", err)
-		}
-		contentStr := normalizeContent(string(content))
-		pages = append(pages, cloudquery_api.PluginDocsPageCreate{
-			Content: contentStr,
-			Name:    strings.TrimSuffix(dirEntry.Name(), fileExt),
-		})
-	}
-
-	if replace {
-		body := cloudquery_api.ReplacePluginVersionDocsJSONRequestBody{
-			Pages: pages,
-		}
-		resp, err := c.ReplacePluginVersionDocsWithResponse(ctx, teamName, cloudquery_api.PluginKind(pluginKind), pluginName, version, body)
-		if err != nil {
-			return fmt.Errorf("failed to upload docs: %w", err)
-		}
-		if resp.HTTPResponse.StatusCode > 299 {
-			return errorFromHTTPResponse(resp.HTTPResponse, resp)
-		}
-	} else {
-		body := cloudquery_api.CreatePluginVersionDocsJSONRequestBody{
-			Pages: pages,
-		}
-		resp, err := c.CreatePluginVersionDocsWithResponse(ctx, teamName, cloudquery_api.PluginKind(pluginKind), pluginName, version, body)
-		if err != nil {
-			return fmt.Errorf("failed to upload docs: %w", err)
-		}
-		if resp.HTTPResponse.StatusCode > 299 {
-			return errorFromHTTPResponse(resp.HTTPResponse, resp)
-		}
-	}
-
-	return nil
-}
-
-func uploadPluginBinary(ctx context.Context, c *cloudquery_api.ClientWithResponses, teamName, pluginName, goos, goarch, localPath string, pkgJSON PackageJSONV1) error {
-	target := goos + "_" + goarch
-	resp, err := c.UploadPluginAssetWithResponse(ctx, teamName, pkgJSON.Kind, pluginName, pkgJSON.Version, target)
-	if err != nil {
-		return fmt.Errorf("failed to upload binary: %w", err)
-	}
-	if resp.HTTPResponse.StatusCode > 299 {
-		msg := fmt.Sprintf("failed to upload binary: %s", resp.HTTPResponse.Status)
-		switch {
-		case resp.JSON403 != nil:
-			msg = fmt.Sprintf("%s: %s", msg, resp.JSON403.Message)
-		case resp.JSON401 != nil:
-			msg = fmt.Sprintf("%s: %s", msg, resp.JSON401.Message)
-		}
-		return fmt.Errorf(msg)
-	}
-	if resp.JSON201 == nil {
-		return fmt.Errorf("upload response is nil, failed to upload binary")
-	}
-	uploadURL := resp.JSON201.Url
-	err = uploadFile(uploadURL, localPath)
-	if err != nil {
-		return fmt.Errorf("failed to upload file: %w", err)
-	}
-	return nil
-}
-
-func readPackageJSON(distDir string) (PackageJSONV1, error) {
-	v := SchemaVersion{}
-	b, err := os.ReadFile(filepath.Join(distDir, "package.json"))
-	if err != nil {
-		return PackageJSONV1{}, err
-	}
-	err = json.Unmarshal(b, &v)
-	if err != nil {
-		return PackageJSONV1{}, err
-	}
-	if v.SchemaVersion != 1 {
-		return PackageJSONV1{}, errors.New("unsupported schema version. This CLI version only supports package.json v1. Try upgrading your CloudQuery CLI version")
-	}
-	pkgJSON := PackageJSONV1{}
-	err = json.Unmarshal(b, &pkgJSON)
-	if err != nil {
-		return PackageJSONV1{}, err
-	}
-	return pkgJSON, nil
 }
