+++ cli/internal/publish/shared.go
@@ -1,0 +1,5 @@
+package publish
+
+type SchemaVersion struct {
+	SchemaVersion int `json:"schema_version"`
+}
